pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watsonx-mlops-model-monitor"
version = "1.0.0"
description = "MLOps Platform with Watsonx Model Monitoring — drift detection, fairness auditing, and Granite-powered alert narratives"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Gabriel Demetrios Lafis", email = "gabriel.lafis@example.com"},
]
keywords = ["watsonx", "mlops", "model-monitoring", "drift-detection", "fairness", "mlflow", "granite"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "ibm-watsonx-ai>=1.1.0",
    "langchain-ibm>=0.3.0",
    "scikit-learn>=1.4.0",
    "mlflow>=2.14.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "streamlit>=1.38.0",
    "prometheus-client>=0.21.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "tenacity>=8.2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
